Move example graphs to their own package.
